backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.4.2
python-dotenv==1.0.0
